smoke_test/requirements.txt
absl-py==1.3.0
